requirements.txt
asgiref==3.8.1
Django==6.0.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
mysqlclient==2.2.7
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
python-dotenv==1.0.1
